[ADD] Creational - The Factory Method Pattern
The purpose of this pattern is to abstract the creation of an object.
The factory method lets a class defer instantiation to subclasses.

This pattern is used in the following use cases:
• A class only knows about abstract classes or interfaces of objects
with whom it has some relationships
• A class wants its subclasses to instantiate the object in order to
benefit of polymorphism mechanism

Participant to this pattern are as follows:
• Product interface: This class contains the definition of our product.
We will define what a card is here.
• Abstract product: This abstract class implements the signature of our
cards and some methods. You'll see that we keep the prototype pattern
that allows us to eventually clone a card. These classes define the
properties of our products.
• Concrete product: This class defines our product; in our example, the
Raid Leader card is a concrete product, such as the Faceless
Manipulator card.
• Concrete creator: This class implements our factory method.