[ADD] Creational - Abstract Factory Pattern
The abstract factory pattern is designed to build objects grouped in
families without having to know the concrete class needed to create the
object.

This pattern is generally used in the following domains:
• A system that uses products needs to stay independent of how these
products are grouped and instantiated
• A system can have several product families that can evolve

The abstract factory pattern has a lot of participants:
• Abstract Factory: This abstract class defines the signature of the
different methods that create our products.
• ConcreteFactory1 and ConcreteFactory2: These are our concrete classes
that implement our methods for each products' families. By knowing the
family and product, the factory is able to create an instance of the
product for that family.
• IProductA and IProductB: These are our interfaces that define our
products that are independent of their family. A family is introduced
in their concrete subclasses.
• ProductA and ProductB: These are the concrete classes that implement
IProductA and IProductB, respectively.